update sql, refactor vo and dto

+++ ems-model/src/main/java/io/clementleetimfu/educationmanagementsystem/pojo/entity/Employee.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -20,7 +21,7 @@ public class Employee {
     private String phone;
     // 1: Class Teacher, 2: Lecturer, 3: Student Affairs Manager, 4: Academic Research Manager, 5: Advisor
     private Integer jobTitle;
-    private Integer salary;
+    private BigInteger salary;
     private String image;
     private LocalDate hireDate;
     private Integer deptId;
